Create PasswordChecker.py
This is a password checking code that helps to get an idea about how many times your password has been used. Your password will be completely safe as it reads the file from your system and apply a hash function to encode the password. Please take a look and try on your machine.